refactor: extract the `EventCard` component
