+++ src/components/EventCard/EventCard-types.ts
