+++ src/components/EventCard/index.ts
@@ -1,0 +1,5 @@
+import EventCard from './EventCard';
+import SkeletonEventCard from './SkeletonEventCard';
+
+export { SkeletonEventCard };
+export default EventCard;
