--- src/components/RecommendActivity/index.ts
@@ -1,3 +1,0 @@
-import RecommendActivity from './RecommendActivity';
-
-export default RecommendActivity;
